jobs/etl_jobs/external/ppg/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in -o requirements.txt
cachetools==5.5.2
    # via google-auth
certifi==2025.8.3
    # via requests
charset-normalizer==3.4.3
    # via requests
click==8.2.1
    # via typer-slim
dateutils==0.6.12
    # via -r requirements.in
db-dtypes==1.4.3
    # via pandas-gbq
duckdb==1.3.2
    # via -r requirements.in
et-xmlfile==2.0.0
    # via openpyxl
google-api-core==2.25.1
    # via
    #   google-api-python-client
    #   google-cloud-bigquery
    #   google-cloud-bigquery-storage
    #   google-cloud-core
    #   google-cloud-secret-manager
    #   google-cloud-storage
    #   pandas-gbq
google-api-python-client==2.185.0
    # via -r requirements.in
google-auth==2.40.3
    # via
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-auth-oauthlib
    #   google-cloud-bigquery
    #   google-cloud-bigquery-storage
    #   google-cloud-core
    #   google-cloud-secret-manager
    #   google-cloud-storage
    #   pandas-gbq
    #   pydata-google-auth
google-auth-httplib2==0.2.0
    # via google-api-python-client
google-auth-oauthlib==1.2.0
    # via
    #   -r requirements.in
    #   pandas-gbq
    #   pydata-google-auth
google-cloud-bigquery==3.36.0
    # via
    #   -r requirements.in
    #   pandas-gbq
google-cloud-bigquery-storage==2.32.0
    # via -r requirements.in
google-cloud-core==2.4.3
    # via
    #   google-cloud-bigquery
    #   google-cloud-storage
google-cloud-secret-manager==2.20.0
    # via -r requirements.in
google-cloud-storage==2.19.0
    # via -r requirements.in
google-crc32c==1.7.1
    # via
    #   google-cloud-storage
    #   google-resumable-media
google-resumable-media==2.7.2
    # via
    #   google-cloud-bigquery
    #   google-cloud-storage
googleapis-common-protos==1.70.0
    # via
    #   google-api-core
    #   grpc-google-iam-v1
    #   grpcio-status
grpc-google-iam-v1==0.14.2
    # via google-cloud-secret-manager
grpcio==1.74.0
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
grpcio-status==1.62.3
    # via google-api-core
httplib2==0.31.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
idna==3.10
    # via requests
logging==0.4.9.6
    # via -r requirements.in
markdown-it-py==4.0.0
    # via rich
mdurl==0.1.2
    # via markdown-it-py
numpy==2.2.6
    # via
    #   -r requirements.in
    #   db-dtypes
    #   pandas
    #   pandas-gbq
oauthlib==3.3.1
    # via requests-oauthlib
openpyxl==3.1.5
    # via -r requirements.in
packaging==25.0
    # via
    #   db-dtypes
    #   google-cloud-bigquery
    #   pandas-gbq
pandas==2.2.2
    # via
    #   -r requirements.in
    #   db-dtypes
    #   pandas-gbq
pandas-gbq==0.23.1
    # via -r requirements.in
proto-plus==1.26.1
    # via
    #   google-api-core
    #   google-cloud-bigquery-storage
    #   google-cloud-secret-manager
protobuf==4.25.8
    # via
    #   google-api-core
    #   google-cloud-bigquery-storage
    #   google-cloud-secret-manager
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
    #   proto-plus
pyarrow==21.0.0
    # via
    #   -r requirements.in
    #   db-dtypes
    #   pandas-gbq
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pydata-google-auth==1.9.1
    # via pandas-gbq
pygments==2.19.2
    # via rich
pyparsing==3.2.5
    # via httplib2
python-dateutil==2.9.0.post0
    # via
    #   -r requirements.in
    #   dateutils
    #   google-cloud-bigquery
    #   pandas
pytz==2025.2
    # via
    #   dateutils
    #   pandas
requests==2.32.5
    # via
    #   google-api-core
    #   google-cloud-bigquery
    #   google-cloud-storage
    #   requests-oauthlib
requests-oauthlib==2.0.0
    # via google-auth-oauthlib
rich==14.1.0
    # via typer-slim
rsa==4.9.1
    # via google-auth
setuptools==80.9.0
    # via
    #   pandas-gbq
    #   pydata-google-auth
shellingham==1.5.4
    # via typer-slim
six==1.17.0
    # via
    #   python-dateutil
    #   treelib
treelib==1.8.0
    # via -r requirements.in
typer==0.12.0
    # via -r requirements.in
typer-cli==0.12.0
    # via typer
typer-slim==0.12.0
    # via
    #   typer
    #   typer-cli
typing-extensions==4.15.0
    # via typer-slim
tzdata==2025.2
    # via pandas
uritemplate==4.2.0
    # via google-api-python-client
urllib3==2.5.0
    # via requests
